Создан файл mixins.py
